requirements.txt
allure-pytest==2.8.18
pytest==6.0.1
pytest-xdist==2.1.0
requests==2.24.0